Add minor changes

+++ src/Services/TizianaTerenzi.Services.Data/CommentsService.cs
@@ -1,11 +1,12 @@
 ﻿namespace TizianaTerenzi.Services.Data
 {
-    using Microsoft.EntityFrameworkCore;
     using System.Linq;
     using System.Threading.Tasks;
 
+    using Microsoft.EntityFrameworkCore;
     using TizianaTerenzi.Data.Common.Repositories;
     using TizianaTerenzi.Data.Models;
+    using TizianaTerenzi.Web.ViewModels.Comments;
 
     public class CommentsService : ICommentsService
     {
@@ -16,18 +17,20 @@
             this.commentsRepository = commentsRepository;
         }
 
-        public async Task CreateAsync(int productId, string userId, string content, int? parentId = null)
+        public async Task<bool> CreateAsync(CreateCommentInputModel inputModel, string userId)
         {
             var comment = new Comment
             {
-                Content = content,
+                Content = inputModel.Content,
                 UserId = userId,
-                ProductId = productId,
-                ParentId = parentId,
+                ProductId = inputModel.ProductId,
+                ParentId = inputModel.ParentId,
             };
 
             await this.commentsRepository.AddAsync(comment);
-            await this.commentsRepository.SaveChangesAsync();
+            var result = await this.commentsRepository.SaveChangesAsync();
+
+            return result > 0;
         }
 
         public async Task DeleteRangeAsync(int productId)
@@ -37,17 +40,20 @@
                 .Where(c => c.ProductId == productId)
                 .ToListAsync();
 
-            this.commentsRepository.DeleteRange(comments);
-            await this.commentsRepository.SaveChangesAsync();
+            if (comments.Any())
+            {
+                this.commentsRepository.DeleteRange(comments);
+                await this.commentsRepository.SaveChangesAsync();
+            }
         }
 
-        public bool IsInProductId(int commentId, int productId)
+        public async Task<bool> IsInProductIdAsync(int commentId, int productId)
         {
-            var commentProductId = this.commentsRepository
+            var commentProductId = await this.commentsRepository
                 .All()
                 .Where(c => c.Id == commentId)
                 .Select(c => c.ProductId)
-                .SingleOrDefault();
+                .SingleOrDefaultAsync();
 
             return commentProductId == productId;
         }
