+++ src/Web/TizianaTerenzi.Web/Areas/Administration/Controllers/ProductsController.cs
@@ -232,6 +232,11 @@
         [HttpPost]
         public async Task<IActionResult> Delete(int id)
         {
+            if (id <= 0)
+            {
+                return this.NotFound();
+            }
+
             try
             {
                 await this.notesService.DeleteProductNotesAsync(id);
